backend/requirements.txt
djangorestframework
markdown
django-filter
pillow
gunicorn
django-cleanup
requests